# docs/openapi/swagger.yaml
components:
  schemas:
...
    access.Item:
      properties:
        builtIn:
          description: BuiltIn is true for items that ship with Warden.
          type: boolean
        credentials:
          description: Credentials are the individual credential entries in this group.
          items:
            $ref: '#/components/schemas/access.Credential'
          type: array
          uniqueItems: false
        description:
          description: Description explains what this access item provides.
          type: string
        id:
          description: |-
            ID is a stable identifier. Built-in items use well-known IDs
            (e.g. "git", "ssh"); user items get generated UUIDs.
          type: string
        label:
          description: Label is the human-readable display name (e.g. "Git Config").
          type: string
        method:
          $ref: '#/components/schemas/access.Method'
      type: object
...
    api.AccessItemListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/api.AccessItemResponse'
          type: array
          uniqueItems: false
      type: object
...
    api.CreateAccessItemRequest:
      properties:
        credentials:
          items:
            $ref: '#/components/schemas/access.Credential'
          type: array
          uniqueItems: false
        description:
          type: string
        label:
          type: string
      type: object
...
    api.UpdateAccessItemRequest:
      properties:
        credentials:
          items:
            $ref: '#/components/schemas/access.Credential'
          type: array
          uniqueItems: false
        description:
          type: string
        label:
          type: string
      type: object
...
    server.apiError:
      properties:
        code:
          $ref: '#/components/schemas/server.ErrorCode'
        error:
          type: string
      type: object
...
info:
  description: Container engine and control plane for running multiple Claude Code
    agents in isolated containers.
  license:
    name: Apache-2.0
    url: https://github.com/thesimonho/warden/blob/main/LICENSE
  title: Warden API
  version: "1.0"
openapi: 3.1.0
paths:
  /api/v1/access:
    get:
      description: |-
        Returns all access items (built-in + user-created) enriched with
        per-credential host detection status. Built-in items that have been
        customized via the DB are returned with the customized configuration.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AccessItemListResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: List access items
      tags:
      - access
    post:
      description: |-
        Creates a new user-defined access item with the given label, description,
        and credential configuration. Returns the created item with a generated ID.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.CreateAccessItemRequest'
                description: Access item configuration
                summary: body
        description: Access item configuration
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/access.Item'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Invalid input (missing label or credentials)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Create access item
      tags:
      - access
  /api/v1/access/{id}:
    delete:
      description: |-
        Deletes a user-defined access item. Built-in items cannot be deleted —
        use the reset endpoint instead.
      parameters:
      - description: Access item ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Cannot delete built-in item
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Delete access item
      tags:
      - access
    get:
      description: |-
        Returns a single access item with detection status. For built-in items,
        returns the DB override if one exists, otherwise the default configuration.
      parameters:
      - description: Access item ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AccessItemResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Item not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Get access item
      tags:
      - access
    put:
      description: |-
        Updates an access item. For built-in items, saves a customized copy to the DB
        (overriding the default). For user items, updates the existing DB row.
        Only provided fields are changed.
      parameters:
      - description: Access item ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.UpdateAccessItemRequest'
                description: Fields to update (all optional)
                summary: body
        description: Fields to update (all optional)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/access.Item'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Invalid input
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Item not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Update access item
      tags:
      - access
  /api/v1/access/{id}/reset:
    post:
      description: |-
        Restores a built-in access item to its default configuration by removing
        any DB override. Only works for built-in items (git, ssh, gpg).
      parameters:
      - description: Built-in access item ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/access.Item'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Not a built-in item
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Reset access item
      tags:
      - access
  /api/v1/access/resolve:
    post:
      description: |-
        Resolves the given access items by checking host sources and computing
        the injections (env vars, mounts) that would be applied to containers.
        Used by the UI "Test" button to preview resolution without creating a container.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.ResolveAccessItemsRequest'
                description: Item IDs to resolve
                summary: body
        description: Item IDs to resolve
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ResolveAccessItemsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Invalid request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Item not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Resolve access items
      tags:
      - access
  /api/v1/audit:
    delete:
      description: |-
        Removes audit events matching the given filters. Supports scoping by
        project, worktree, and time range.
      parameters:
      - description: Filter by project ID
        in: query
        name: projectId
        schema:
          type: string
      - description: Filter by worktree ID
        in: query
        name: worktree
        schema:
          type: string
      - description: Delete entries after this timestamp (RFC3339)
        in: query
        name: since
        schema:
          type: string
      - description: Delete entries before this timestamp (RFC3339)
        in: query
        name: until
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  deleted:
                    type: integer
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Delete audit events
      tags:
      - audit
    get:
      description: |-
        Returns audit-relevant events (sessions, tool uses, prompts, lifecycle)
        with optional filtering by project, worktree, category, and time range.
      parameters:
      - description: Filter by project ID
        in: query
        name: projectId
        schema:
          type: string
      - description: Filter by worktree ID
        in: query
        name: worktree
        schema:
          type: string
      - description: Filter by category (session, agent, prompt, config, budget, system,
          debug)
        in: query
        name: category
        schema:
          type: string
      - description: Filter by source (agent, backend, frontend, container)
        in: query
        name: source
        schema:
          type: string
      - description: Filter by level (info, warning, error)
        in: query
        name: level
        schema:
          type: string
      - description: Filter entries after this timestamp (RFC3339)
        in: query
        name: since
        schema:
          type: string
      - description: Filter entries before this timestamp (RFC3339)
        in: query
        name: until
        schema:
          type: string
      - description: Maximum entries to return (default 10000)
        in: query
        name: limit
        schema:
          type: integer
      - description: Number of entries to skip
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.AuditEntry'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Get audit log
      tags:
      - audit
    post:
      description: |-
        Writes a custom audit event. Integrators can set source, project scope,
        and structured data. Source defaults to "external" if omitted.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.PostAuditEventRequest'
                description: Audit event to record
                summary: body
        description: Audit event to record
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Post audit event
      tags:
      - audit
  /api/v1/audit/export:
    get:
      description: Downloads audit events as CSV or JSON for compliance review.
      parameters:
      - description: 'Export format: csv or json (default json)'
        in: query
        name: format
        schema:
          type: string
      - description: Filter by project ID
        in: query
        name: projectId
        schema:
          type: string
      - description: Filter by worktree ID
        in: query
        name: worktree
        schema:
          type: string
      - description: Filter by category (session, agent, prompt, config, budget, system,
          debug)
        in: query
        name: category
        schema:
          type: string
      - description: Filter entries after this timestamp (RFC3339)
        in: query
        name: since
        schema:
          type: string
      - description: Filter entries before this timestamp (RFC3339)
        in: query
        name: until
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Export audit log
      tags:
      - audit
  /api/v1/audit/projects:
    get:
      description: Returns distinct project IDs that have audit events recorded.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: string
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: List audit projects
      tags:
      - audit
  /api/v1/audit/summary:
    get:
      description: |-
        Returns aggregate statistics for audit events including session counts,
        tool usage, cost totals, and top tools.
      parameters:
      - description: Filter by project ID
        in: query
        name: projectId
        schema:
          type: string
      - description: Filter by worktree ID
        in: query
        name: worktree
        schema:
          type: string
      - description: Filter entries after this timestamp (RFC3339)
        in: query
        name: since
        schema:
          type: string
      - description: Filter entries before this timestamp (RFC3339)
        in: query
        name: until
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AuditSummary'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Get audit summary
      tags:
      - audit
...
  /api/v1/events:
    get:
      description: |-
        Opens a Server-Sent Events stream for real-time updates. Optionally filter
        by projectId and agentType to receive events for a single project.
        Event types: worktree_state, project_state, worktree_list_changed,
        budget_exceeded, budget_container_stopped, heartbeat, server_shutdown,
        runtime_status, agent_status.
      parameters:
      - description: Filter events to this project ID
        in: query
        name: projectId
        schema:
          type: string
      - description: Filter events to this agent type
        in: query
        name: agentType
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: SSE event stream
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Event streaming not configured
      summary: Subscribe to events (SSE)
      tags:
      - streaming
  /api/v1/filesystem/directories:
    get:
      description: |-
        Returns filesystem entries at the given path. By default only directories
        are returned. Pass mode=file to include files alongside directories.
      parameters:
      - description: Absolute path to list entries in
        in: query
        name: path
        required: true
        schema:
          type: string
      - description: 'Browse mode: omit for directories only, ''file'' for directories
          and files'
        in: query
        name: mode
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.DirEntry'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: List filesystem entries
      tags:
      - host
  /api/v1/filesystem/reveal:
    post:
      description: Opens the given host directory in the system file manager (Finder,
        Nautilus, Explorer).
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/server.revealRequest'
                description: Directory path to reveal
                summary: body
        description: Directory path to reveal
        required: true
      responses:
        "204":
          description: Directory opened
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Path not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Reveal in file manager
      tags:
      - host
  /api/v1/focus:
    get:
      description: Returns the aggregated viewer focus state across all connected
        clients.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.FocusState'
          description: OK
      summary: Get focus state
      tags:
      - focus
    post:
      description: |-
        Reports which project and worktrees a client is actively viewing.
        Used by the system tray to suppress desktop notifications for focused projects.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.FocusRequest'
                description: Focus state report
                summary: body
        description: Focus state report
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Bad Request
      summary: Report focus state
      tags:
      - focus
...
  /api/v1/projects:
    get:
      description: |-
        Returns all configured projects enriched with live container state,
        Claude status, worktree counts, and cost data.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.ProjectResponse'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: List projects
      tags:
      - projects
    post:
      description: |-
        Registers a host directory as a Warden project. Optionally creates a
        container in the same request by including a "container" field. If
        container creation fails, the project is cleaned up automatically.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.AddProjectRequest'
                description: Project details (with optional container config)
                summary: body
        description: Project details (with optional container config)
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AddProjectResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Container name already in use
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Add project
      tags:
      - projects
...
  /api/v1/projects/{projectId}/{agentType}/audit:
    delete:
      description: Removes all audit events for the given project.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: integer
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.apiError'
          description: Internal Server Error
      summary: Purge project audit
      tags:
      - projects
...
servers:
- description: Local development server
  url: http://localhost:8090